+++ src/Lifecycle/DelegatingLifecycle.php
@@ -37,6 +37,17 @@ final class DelegatingLifecycle implements MessageLifecycle, \Countable
 
     public static function fromArray(array $lifecycles) : self
     {
+        @trigger_error(sprintf(
+            '%s is deprecated as of version 5.0 and will be removed in 6.0, use %s::fromIterable instead',
+            __METHOD__,
+            __CLASS__
+        ), E_USER_DEPRECATED);
+
+        return new self(...$lifecycles);
+    }
+
+    public static function fromIterable(iterable $lifecycles) : self
+    {
         return new self(...$lifecycles);
     }
 
@@ -63,10 +74,20 @@ final class DelegatingLifecycle implements MessageLifecycle, \Countable
     /**
      * {@inheritdoc}
      */
-    public function failed(Message $message, Consumer $consumer, bool $isRetrying)
+    public function retrying(Message $message, Consumer $consumer)
     {
-        $this->apply(function (MessageLifecycle $ml) use ($message, $consumer, $isRetrying) {
-            $ml->failed($message, $consumer, $isRetrying);
+        $this->apply(function (MessageLifecycle $ml) use ($message, $consumer) {
+            $ml->retrying($message, $consumer);
+        });
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function failed(Message $message, Consumer $consumer)
+    {
+        $this->apply(function (MessageLifecycle $ml) use ($message, $consumer) {
+            $ml->failed($message, $consumer);
         });
     }