+++ src/Lifecycle/MappingLifecycle.php
@@ -76,9 +76,17 @@ final class MappingLifecycle implements MessageLifecycle
     /**
      * {@inheritdoc}
      */
-    public function failed(Message $message, Consumer $consumer, bool $isRetrying)
+    public function retrying(Message $message, Consumer $consumer)
     {
-        $this->lifecycleFor($message)->failed($message, $consumer, $isRetrying);
+        $this->lifecycleFor($message)->retrying($message, $consumer);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function failed(Message $message, Consumer $consumer)
+    {
+        $this->lifecycleFor($message)->failed($message, $consumer);
     }
 
     /**